fix examples. compiles already, but does not link yet
git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@6051 ce417995-dfc9-0310-95a0-acdaff106893

+++ tests/buffers_test.cpp
@@ -341,7 +341,7 @@ void BuffersTest::testBufLockFree()
     BufferBase::size_type sz = 10;
     CPPUNIT_ASSERT( lockfree->write( *c ) );
     CPPUNIT_ASSERT( lockfree->write( *d ) );
-    CPPUNIT_ASSERT( lockfree->write( v ) == v.size() );
+    CPPUNIT_ASSERT( lockfree->write( v ) == (int)v.size() );
     CPPUNIT_ASSERT( lockfree->write( *c ) );
     CPPUNIT_ASSERT( lockfree->write( *d ) );
     CPPUNIT_ASSERT( lockfree->write( v ) == 1 );
